+++ scripts/logging/dictionary/dictionary_parser/log_database.py
