+++ scripts/logging/dictionary/dictionary_parser/log_parser.py
