+++ scripts/logging/dictionary/dictionary_parser/log_parser_v1.py
