+++ scripts/logging/dictionary/dictionary_parser/utils.py
